halv1/requirements.txt
requests>=2.31.0
aiohttp>=3.9.0
python-telegram-bot>=20.0
PyYAML>=6.0.1
numpy>=1.26.0
Telethon>=1.34.0
faiss-cpu>=1.8.0
pandas>=2.0.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
pandas-stubs>=2.3.0
types-PyYAML>=6.0.12
pydantic>=2.0.0
jmespath>=1.0.1
duckduckgo-search>=8.1.1
ddgs>=4.0.0
feedparser>=6.0.0
# Async backend: asyncio only
hnswlib>=0.7
watchdog>=3.0.0
prometheus_client>=0.19.0
fastapi>=0.104.0
mypy>=1.10.0
uvicorn>=0.24.0
Jinja2>=3.1.0
mcp>=1.14.1
asyncpg>=0.29.0
structlog>=24.1.0
